Storing login information in SQL database
Write a program that takes email and password as input and stores them in a database as username and password. The code written should be able to enter the information as input and store and record the information in a database file.
It is essential to follow the email format as input. Passwords also include numbers and strings like "expression@string.string".